Added structure of controller/driver classes
