+++ src/main/java/EventManager.java
@@ -1,0 +1,5 @@
+package main.java;
+
+public interface EventManager {
+
+}
